Reorganize plugins folder
